Don't `import functorch` inside inductor
We have some internal problems with the functorch target where it is not
automatically included with the torch target. This PR is a quick
short-term fix that excises the dependency on `import functorch` from
PyTorch so that you can use `torch.compile` without needing to be able
to `import functorch`.

Test Plan:
- CI

ghstack-source-id: 1ac8d7397eb6db03b049d7112a6f348c2da2a426
Pull Request resolved: https://github.com/pytorch/pytorch/pull/101802