Add newline before minified repro path
Minor QOL change.  This log message is pushed into my history by the
backtrace, which is a pain because if I tab up in tmux I can no longer
paste it without line breaks.  This makes it more convenient to use tmux
copy mode to get only the file (as I get the entire line this way.)

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: 53964b840c47508ab55b3049282a1231be145db7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103083