[FSDP] Fix 2D fully sharded state_dict regression.
The #100856 PR changed how we load a fully sharded state dict.

Originally, we'd pick the unsharded params from the state_dict,
apply 2D transform and gather them. Then we'd use those tensors
to create a flat param tensor and put that in the state_dict.

The code in module::load_state_dict would then load the flat param
just fine.

Then #100856 changed to keep the state_dict with the unsharded params and
unshard the model as well.

The problem is that the extension peeled out the nested sharded tensor but
was not given an opportunity to unpeel back into a DTensor.

The fix is to just do that, convert the gathered tensor into the DTensor
shard it actually is.

Fixes #101162

ghstack-source-id: 4d2f1dd640da571fc79b68d7b2495d9f2673f5db
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103214